[settings] update generated settings helper
