+++ libfreerdp/common/CMakeLists.txt
@@ -23,6 +23,7 @@ set(${MODULE_PREFIX}_SRCS
 	settings.c
 	settings_getters.c
 	settings_str.c
+	settings_str.h
 	assistance.c)
 
 freerdp_module_add(${${MODULE_PREFIX}_SRCS})
