+++ libfreerdp/common/settings_getters.c
@@ -1,4 +1,5 @@
 /* Generated by  */
+
 #include "../core/settings.h"
 
 #include <winpr/assert.h>
@@ -1534,8 +1535,8 @@ INT16 freerdp_settings_get_int16(const rdpSettings* settings, FreeRDP_Settings_K
 	{
 		default:
 			WLog_ERR(TAG, "Invalid key index %" PRIuz " [%s|%s]", id,
-			         freerdp_settings_get_name_for_key(id),
-			         freerdp_settings_get_type_name_for_key(id));
+			         freerdp_settings_get_name_for_key((size_t)id),
+			         freerdp_settings_get_type_name_for_key((size_t)id));
 			return 0;
 	}
 }
